Docs: keep the playground tests off built packages
