+++ docs/app/playground/evaluate.ts
@@ -1,5 +1,4 @@
-import * as React from "react";
-import { createElement, type ElementType, type ReactNode } from "react";
+import type * as React from "react";
 import { transform } from "sucrase";
 import * as primitives from "takumi-pdf/primitives";
 import * as z from "zod/mini";
@@ -39,20 +38,3 @@ export function evaluateCodeExports(code: string, react: typeof React) {
 
   return exportsSchema.parse(exports);
 }
-
-// Mirror Takumi's `tw` into `className` (keeping `tw`) so one evaluated tree
-// serves both the Takumi render (reads `tw`) and the browser preview (reads `class`).
-function mirrorTw<P>(props: P): P {
-  if (!props || typeof props !== "object" || !("tw" in props)) return props;
-  const { tw, className, class: klass } = props as Record<string, unknown>;
-  return { ...props, className: [className ?? klass, tw].filter(Boolean).join(" ") };
-}
-
-export const renderReact: typeof React = {
-  ...React,
-  createElement: ((
-    type: ElementType,
-    props: Record<string, unknown> | null,
-    ...children: ReactNode[]
-  ) => createElement(type, mirrorTw(props), ...children)) as typeof createElement,
-};
