+++ docs/app/playground/worker.ts
@@ -10,7 +10,8 @@ import pdfWasm from "takumi-pdf/wasm-url";
 import initPdf, { PdfRenderer } from "takumi-pdf/no-init";
 import type { JSXElementConstructor } from "react";
 import { renderToStaticMarkup } from "react-dom/server";
-import { evaluateCodeExports, renderReact } from "./evaluate";
+import { evaluateCodeExports } from "./evaluate";
+import { renderReact } from "./render-react";
 import { FONT_FAMILIES } from "./fonts";
 import { inspectPdf } from "./inspect-pdf";
 import { keyframesToCss } from "./preview-css";
